+++ inference_engine/inferer.cpp
@@ -3,12 +3,13 @@
 
 namespace inference_engine {
 namespace inferer {
-std::pair<int, int> calculate_conv_matrix_dims(int h, int w, int k, int pad,
-                                               int stride) {
-  std::pair<int, int> result;
-  result.first = static_cast<int>(floor((h - k + 2 * pad) / float(stride)) + 1);
+std::pair<long, long> calculate_conv_matrix_dims(long h, long w, long k,
+                                                 long pad, long stride) {
+  std::pair<long, long> result;
+  result.first =
+      static_cast<long>(floor((h - k + 2 * pad) / float(stride)) + 1);
   result.second =
-      static_cast<int>(floor((w - k + 2 * pad) / float(stride)) + 1);
+      static_cast<long>(floor((w - k + 2 * pad) / float(stride)) + 1);
 
   return result;
 }
